[RELEASE] version 2.1.1

+++ src/main/java/com/cashfree/pg/model/CFLink.java
@@ -40,7 +40,7 @@ public class CFLink {
     public static final String SERIALIZED_NAME_CF_LINK_ID = "cf_link_id";
 
     @SerializedName(SERIALIZED_NAME_CF_LINK_ID)
-    private Integer cfLinkId;
+    private long cfLinkId;
 
     public static final String SERIALIZED_NAME_LINK_ID = "link_id";
 
@@ -125,7 +125,7 @@ public class CFLink {
 
     public CFLink() {}
 
-    public CFLink cfLinkId(Integer cfLinkId) {
+    public CFLink cfLinkId(long cfLinkId) {
 
         this.cfLinkId = cfLinkId;
         return this;
@@ -137,11 +137,11 @@ public class CFLink {
      * @return cfLinkId
      */
     @javax.annotation.Nullable
-    public Integer getCfLinkId() {
+    public long getCfLinkId() {
         return cfLinkId;
     }
 
-    public void setCfLinkId(Integer cfLinkId) {
+    public void setCfLinkId(long cfLinkId) {
         this.cfLinkId = cfLinkId;
     }
